+++ src/tdd/state-store/constants.js
@@ -1,0 +1,6 @@
+export let STATE_METADATA_KEYS = {
+  baseline: 'baseline_metadata',
+  hotspot: 'hotspot_metadata',
+  region: 'region_metadata',
+  baselineBuild: 'baseline_build_metadata',
+};
